Changes in music cogs
v0.0.120a
Fixed an issue where a maximum of 600 tracks were uploaded to the player queue.
The music player commands are added in commands.json
Sorted imports in music and radio cogs
The MUSIC_COMMANDS_CHANNEL variable was moved to constants.py
Updated Lavalink.jar
The privileged user now requires administrator permissions
When forming a request to delete a track from radio playlist, a link to the track will be added in the database along with the track name.